Added command assertions and slightly cleaned up the codebase  (#47)
- added command assertions to satisfy existing unit tests
- fixed some code issues according to linting rules

+++ Todo.Backend/Commands/DeleteTodoItemCommandHandler.cs
@@ -3,6 +3,7 @@ using System.Threading;
 using System.Threading.Tasks;
 using Todo.Backend.Commands;
 using Todo.Backend.Contract.Repository;
+using Todo.Backend.Exceptions;
 
 namespace backend.Commands
 {
@@ -11,7 +12,7 @@ namespace backend.Commands
         // TODO: Currently, read and write persistence as well as models are equal, so using the read model is perfectly
         //       fine for the first iteration. However, commands should NEVER use read model persistence to rehydrate the
         //       model for change actions. Later, we need dedicated rehydration logic for the write model.
-        private ITodoReadRepository _readRepository;
+        private readonly ITodoReadRepository _readRepository;
 
         public DeleteTodoItemCommandHandler(
             ITodoWriteRepository repository,
@@ -20,9 +21,16 @@ namespace backend.Commands
             _readRepository = readRepository;
         }
 
-        public async override Task<Unit> Handle(DeleteTodoItemCommand request, CancellationToken cancellationToken)
+        public override async Task<Unit> Handle(DeleteTodoItemCommand request, CancellationToken cancellationToken)
         {
+            AssertValidGuid(request.Id);
+            
             var item = await _readRepository.GetItemById(request.Id, cancellationToken);
+
+            if(item == null)
+            {
+                throw new ItemNotFoundException(request.Id);
+            }
             await Repository.DeleteAsync(item, cancellationToken);
             return Unit.Value;
         }
